chore(cdk): split missing sub-packages (#942)
