+++ apps/demos/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk';
+import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk/render-strategies';
 import { AppComponent, AppComponentModule } from './app-component';
 import { ENVIRONMENT_SETTINGS } from './shared/environment.token';
 import { environment } from '../environments/environment';
