+++ apps/demos/src/app/shared/debug-helper/strategy-select/strategy-select/strategy-select.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, Output } from '@angular/core';
 import { RxState } from '@rx-angular/state';
-import { RxStrategyCredentials, RxStrategyProvider } from '@rx-angular/cdk';
+import { RxStrategyProvider } from '@rx-angular/cdk';
 import { delay, map, skip } from 'rxjs/operators';
 import { EventEmitter } from '@angular/core';
