+++ libs/cdk/coalescing/src/lib/index.ts
@@ -1,3 +1,3 @@
 export { RxCoalescingOptions, coalescingObj } from './model';
 export { coalesceWith } from './coalesceWith';
-export { coalescingManager } from './coalescingManager';
+export { coalescingManager, CoalescingManager } from './coalescingManager';
