+++ libs/cdk/render-strategies/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/package.schema.json",
+  "lib": {
+    "entryFile": "src/index.ts",
+    "flatModuleFile": "cdk-render-strategies",
+    "umdId": "rx-angular.cdk-render-strategies"
+  }
+}
