+++ libs/cdk/render-strategies/src/index.ts
@@ -1,0 +1,20 @@
+export { RxStrategyProvider } from './lib/strategy-provider.service';
+export { ScheduleOnStrategyOptions } from './lib/model';
+export {
+  RX_CONCURRENT_STRATEGIES,
+  RxConcurrentStrategies,
+} from './lib/concurrent-strategies';
+export { RX_NATIVE_STRATEGIES, RxNativeStrategies } from './lib/native-strategies';
+export { onStrategy } from './lib/onStrategy';
+export { strategyHandling } from './lib/strategy-handling';
+export { RxStrategies } from './lib/model';
+export { RxStrategyNames } from './lib/model';
+export { RxDefaultStrategyNames } from './lib/model';
+export { RxConcurrentStrategyNames } from './lib/model';
+export { RxNativeStrategyNames } from './lib/model';
+export { RxCustomStrategyCredentials } from './lib/model';
+export { RxStrategyCredentials } from './lib/model';
+export { RxRenderBehavior } from './lib/model';
+export { RxRenderWork } from './lib/model';
+export { RX_ANGULAR_CONFIG } from './lib/config';
+export { RxAngularConfig } from './lib/config';
