+++ libs/cdk/render-strategies/src/lib/onStrategy.ts
@@ -1,13 +1,7 @@
 import { RxCoalescingOptions } from '@rx-angular/cdk/coalescing';
 import { switchMap, take } from 'rxjs/operators';
-import {
-  RxRenderWork,
-  RxStrategyCredentials,
-} from '../model';
 import { Observable, of, throwError } from 'rxjs';
-import {
-  RxRenderErrorFactory,
-} from '../template-management/render-error';
+import { RxRenderWork, RxStrategyCredentials } from './model';
 
 /**
  * @internal
@@ -16,7 +10,6 @@ import {
  * @param strategy
  * @param workFactory
  * @param options
- * @param errorFactory
  */
 export function onStrategy<T>(
   value: T,
@@ -26,8 +19,7 @@ export function onStrategy<T>(
     work: RxRenderWork,
     options: RxCoalescingOptions
   ) => void,
-  options: RxCoalescingOptions = {},
-  errorFactory: RxRenderErrorFactory<T, any> = (e, v) => [e, v]
+  options: RxCoalescingOptions = {}
 ): Observable<T> {
   let error: Error;
   return new Observable<T>(subscriber => {
@@ -41,7 +33,7 @@ export function onStrategy<T>(
       }
     }, options.scope || {}),
     switchMap(() =>
-      error ? throwError(errorFactory(error, value)) : of(value)
+      error ? throwError([error, value]) : of(value)
     ),
     take(1)
   );
