+++ libs/cdk/spec/template-management/list-manager.spec.ts
@@ -15,13 +15,12 @@ import {
 import { TestBed } from '@angular/core/testing';
 // tslint:disable-next-line:nx-enforce-module-boundaries
 import {
-  createListTemplateManager,
-  RX_ANGULAR_CONFIG,
+  createListTemplateManager, RX_ANGULAR_CONFIG,
   RxDefaultListViewContext,
   RxListManager,
   RxListViewComputedContext,
   RxListViewContext,
-  RxStrategyProvider,
+  RxStrategyProvider
 } from '@rx-angular/cdk';
 // tslint:disable-next-line:nx-enforce-module-boundaries
 import { mockConsole } from '@test-helpers';
