+++ libs/cdk/src/index.ts
@@ -1,14 +1,10 @@
-export * from './lib/model';
-export * from './lib/utils';
-export * from './lib/render-strategies';
-export * from './lib/template-management';
-export * from './lib/cdk-config';
-export * from './lib/state';
-
 /* @todo: remove the following backward compatibility barrel exports when we have a migration script. */
 export * from '@rx-angular/cdk/coalescing';
 export * from '@rx-angular/cdk/coercing';
 export * from '@rx-angular/cdk/internals/scheduler';
+export * from '@rx-angular/cdk/render-strategies';
+export * from '@rx-angular/cdk/state';
+export * from '@rx-angular/cdk/template';
 export * from '@rx-angular/cdk/notifications';
 export * from '@rx-angular/cdk/zone-less';
 export * from '@rx-angular/cdk/zone-configurations';
