--- libs/cdk/src/lib/render-strategies/model.ts
@@ -1,7 +1,0 @@
-import { NgZone } from '@angular/core';
-
-export interface ScheduleOnStrategyOptions {
-  scope?: {};
-  strategy?: string;
-  patchZone?: false | NgZone;
-}
