--- libs/cdk/src/lib/state/index.ts
@@ -1,2 +1,0 @@
-export { ObservableAccumulation, ObservableMap } from './model';
-export { accumulateObservables } from './accumulateObservables';
