--- libs/cdk/src/lib/utils/index.ts
@@ -1,2 +1,0 @@
-export { onStrategy } from './onStrategy';
-export { strategyHandling } from './strategy-handling';
