+++ libs/cdk/state/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/package.schema.json",
+  "lib": {
+    "entryFile": "src/index.ts",
+    "flatModuleFile": "cdk-state",
+    "umdId": "rx-angular.cdk-state"
+  }
+}
