+++ libs/cdk/state/src/index.ts
@@ -1,0 +1,2 @@
+export { ObservableAccumulation, ObservableMap } from './lib/model';
+export { accumulateObservables } from './lib/accumulateObservables';
