+++ libs/cdk/template/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/package.schema.json",
+  "lib": {
+    "entryFile": "src/index.ts",
+    "flatModuleFile": "cdk-template",
+    "umdId": "rx-angular.cdk-template"
+  }
+}
