+++ libs/cdk/template/src/lib/list-view-handler.ts
