+++ libs/template/let/src/lib/tests/let.directive.error.spec.ts
@@ -4,13 +4,13 @@ import {
   TemplateRef,
   ViewContainerRef,
 } from '@angular/core';
+import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk/render-strategies';
 import { Observable, of, throwError } from 'rxjs';
 import { TestBed, waitForAsync } from '@angular/core/testing';
 import { LetDirective } from '../let.directive';
 import { MockChangeDetectorRef } from './fixtures';
 // tslint:disable-next-line:nx-enforce-module-boundaries
 import { mockConsole } from '@test-helpers';
-import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk';
 
 @Component({
   template: `
