+++ libs/template/let/src/lib/tests/let.directive.rendered.spec.ts
@@ -5,7 +5,7 @@ import {
   ViewContainerRef,
 } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk';
+import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk/render-strategies';
 // tslint:disable-next-line:nx-enforce-module-boundaries
 import { mockConsole } from '@test-helpers';
 import { BehaviorSubject, Observable, of, Subject } from 'rxjs';
