+++ libs/template/let/src/lib/tests/let.directive.template-binding.no-error.spec.ts
@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk/render-strategies';
 import { LetDirective } from '../let.directive';
 import { Observable, of, Subject } from 'rxjs';
 // tslint:disable-next-line:nx-enforce-module-boundaries
 import { mockConsole } from '@test-helpers';
-import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk';
 
 @Component({
   template: `
