+++ libs/template/push/src/lib/tests/push.pipe.spec.ts
@@ -1,4 +1,5 @@
-import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk';
+
+import { RX_ANGULAR_CONFIG } from '@rx-angular/cdk/render-strategies';
 import { PushPipe } from '../push.pipe';
 import { TestBed } from '@angular/core/testing';
 import { ChangeDetectorRef, Component } from '@angular/core';
