+++ libs/template/tsconfig.lib.json
@@ -7,11 +7,10 @@
     "importHelpers": true,
     "lib": ["dom", "es2018"],
     "paths": {
-      "@rx-angular/cdk": ["dist/libs/cdk"],
       "@rx-angular/cdk/zone-less": ["dist/libs/cdk/zone-less"],
-      "@rx-angular/cdk/zone-configurations": [
-        "dist/libs/cdk/zone-configurations"
-      ],
+      "@rx-angular/cdk/zone-configurations": ["dist/libs/cdk/zone-configurations"],
+      "@rx-angular/cdk/render-strategies": ["dist/libs/cdk/render-strategies"],
+      "@rx-angular/cdk/template": ["dist/libs/cdk/template"],
       "@rx-angular/cdk/coercing": ["dist/libs/cdk/coercing"],
       "@rx-angular/cdk/notifications": ["dist/libs/cdk/notifications"]
     }
