+++ tsconfig.base.json
@@ -22,6 +22,15 @@
         "libs/cdk/internals/scheduler/src/index.ts"
       ],
       "@rx-angular/cdk/notifications": ["libs/cdk/notifications/src/index.ts"],
+      "@rx-angular/cdk/render-strategies": [
+        "libs/cdk/render-strategies/src/index.ts"
+      ],
+      "@rx-angular/cdk/state": [
+        "libs/cdk/state/src/index.ts"
+      ],
+      "@rx-angular/cdk/template": [
+        "libs/cdk/template/src/index.ts"
+      ],
       "@rx-angular/cdk/zone-configurations": [
         "libs/cdk/zone-configurations/src/index.ts"
       ],
